api/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose[cryptography]
redis
sqlalchemy
asyncpg
chromadb
llama-cpp-python
